+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.6.0-alpha4
+
+* Publish the raw content-negotiation results into `FormatAndCharset` too
+* Removed helpers `m/get-negotiated-request-content-type` and `m/get-negotiated-response-content-type`
+* Added helpers `m/get-request-format-and-charset` and `get-response-format-and-charset`
+
+```clj
+(require '[muuntaja.middleware :as middleware])
+(require '[muuntaja.core :as m])
+
+(->> {:headers {"accept" "cheese/cake"
+                "accept-charset" "cheese-16"}}
+     ((middleware/wrap-format identity))
+     (m/get-response-format-and-charset))
+;#FormatAndCharset{:format "application/json"
+;                  :charset "utf-8"
+;                  :raw-format "cheese/cake"
+;                  :raw-charset "cheese-16"}
+```
+
 ## 0.6.0-alpha3
 
 * If `:body-params` is set in request, don't try to decode request body.